requirements.txt
psutil>=5.9
PyYAML>=6.0
requests>=2.31
pystray>=0.19
Pillow>=10.0